Moved LogMessage to a struct file Added interface for an external logger Changed Logger to have an external logger Added ability to change Logger output mode
