+++ src/Code/Engine/Interfaces/IExternalLogger.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "DataStructs/LogMessage.h"
+
+namespace EngineECS {
+	class IExternalLogger {
+	public:
+		virtual void WriteLog(LogMessage& message) = 0;
+	};
+}
