+++ src/Code/Engine/Tools/Logger.h
@@ -3,6 +3,8 @@
 #include <string>
 #include <queue>
 #include <iostream>
+#include "DataStructs/LogMessage.h"
+#include <Interfaces\IExternalLogger.h>
 
 using std::string;
 using std::queue;
@@ -17,30 +19,13 @@ namespace EngineECS {
 		External // If mode changes to external and link is broken, default to console
 	};
 
-	enum class LogLevel {
-		Info,
-		Warning,
-		Error
-	};
-
-	struct LogMessage {
-		LogLevel _logLevel;
-		unsigned int _index;
-		string _message;
-
-		LogMessage(LogLevel logLevel, unsigned int index, string message) :
-			_logLevel(logLevel),
-			_index(index),
-			_message(message) {
-		}
-	};
-
 	class Logger {
 	private:
 		static Logger* _instance;
 		Logger();
 		~Logger();
 
+		IExternalLogger* _externalLogger;
 		LoggingDestination _loggingDestination;
 		unsigned int _logIndex; // Used to order the messages
 		unsigned int _maxLogCount;
@@ -60,6 +45,9 @@ namespace EngineECS {
 		
 		void LogError(const char* message);
 		void LogError(string& message);
+
+		void SetLoggingDestination(LoggingDestination loggingDestination);
+		void SetExternalLogger(IExternalLogger* externalLogger);
 	};
 }
